src/main/java/com/dscommerce/dscommerce/services/CategoryService.java
package com.dscommerce.dscommerce.services;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import com.dscommerce.dscommerce.dto.CategoryDTO;
import com.dscommerce.dscommerce.entities.Category;
import com.dscommerce.dscommerce.repositories.CategoryRepository;

@Service
public class CategoryService {
    
    @Autowired
    private CategoryRepository repository;

    @Transactional(readOnly = true)
    public List<CategoryDTO> findAll() {
        List<Category> result = repository.findAll();
        return result.stream().map(CategoryDTO::new).toList();
    }
}
